Refactor some stuff which severely annoyed me
