+++ src/main/java/supersymmetry/api/recipes/builders/SinteringRecipeBuilder.java
@@ -1,13 +1,11 @@
-package supersymmetry.api.recipe.builders;
+package supersymmetry.api.recipes.builders;
 
 import gregtech.api.recipes.Recipe;
 import gregtech.api.recipes.RecipeBuilder;
 import gregtech.api.recipes.RecipeMap;
-import gregtech.api.util.EnumValidationResult;
-import gregtech.api.util.GTLog;
 import org.jetbrains.annotations.NotNull;
-import supersymmetry.api.recipe.properties.CoilingCoilTemperatureProperty;
-import supersymmetry.api.recipe.properties.SinterProperty;
+import supersymmetry.api.recipes.properties.CoilingCoilTemperatureProperty;
+import supersymmetry.api.recipes.properties.SinterProperty;
 
 public class SinteringRecipeBuilder extends RecipeBuilder<SinteringRecipeBuilder> {
