+++ src/main/java/supersymmetry/common/blocks/BlockCoolingCoil.java
@@ -1,4 +1,4 @@
-package supersymmetry.api.block;
+package supersymmetry.common.blocks;
 
 import gregtech.api.block.VariantActiveBlock;
 import gregtech.api.block.VariantItemBlock;
