+++ src/main/java/supersymmetry/common/blocks/SusyBlockStoneSmooth.java
@@ -1,4 +1,4 @@
-package supersymmetry.api.block;
+package supersymmetry.common.blocks;
 
 import javax.annotation.Nonnull;
