+++ src/main/java/supersymmetry/common/metatileentities/single/steam/MetaTileEntitySteamVulcanizingPress.java
@@ -17,7 +17,7 @@ import net.minecraftforge.fluids.IFluidTank;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 import net.minecraftforge.items.IItemHandlerModifiable;
-import supersymmetry.api.recipe.SuSyRecipeMaps;
+import supersymmetry.api.recipes.SuSyRecipeMaps;
 import supersymmetry.client.renderer.textures.SusyTextures;
