+++ src/main/java/supersymmetry/loaders/recipes/CoagulationRecipes.java
@@ -1,6 +1,7 @@
-package supersymmetry.api.recipe;
+package supersymmetry.loaders.recipes;
 
 import gregtech.api.unification.ore.OrePrefix;
+import supersymmetry.api.recipes.SuSyRecipeMaps;
 import supersymmetry.common.materials.SusyMaterials;
 
 public class CoagulationRecipes {
