+++ src/main/java/supersymmetry/loaders/recipes/FridgeRecipes.java
@@ -1,6 +1,7 @@
-package supersymmetry.api.recipe;
+package supersymmetry.loaders.recipes;
 
 import net.minecraft.init.Items;
+import supersymmetry.api.recipes.SuSyRecipeMaps;
 import supersymmetry.common.metatileentities.SuSyMetaTileEntities;
 
 public class FridgeRecipes {
