+++ src/main/java/supersymmetry/loaders/recipes/OreRecipeHandler.java
@@ -1,0 +1,9 @@
+package supersymmetry.loaders.recipes;
+
+public class OreRecipeHandler {
+    public static void init(){
+
+
+    }
+
+}
