+++ src/main/java/supersymmetry/loaders/recipes/VulcanizationRecipes.java
@@ -1,8 +1,9 @@
-package supersymmetry.api.recipe;
+package supersymmetry.loaders.recipes;
 
 import gregtech.api.unification.material.Materials;
 import gregtech.api.unification.ore.OrePrefix;
 import net.minecraft.init.Items;
+import supersymmetry.api.recipes.SuSyRecipeMaps;
 import supersymmetry.common.materials.SusyMaterials;
 
 public class VulcanizationRecipes {
